% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{ddirichlet}
\alias{ddirichlet}
\alias{rdirichlet}
\title{Dirichlet distribution}
\usage{
ddirichlet(x, alpha, log = FALSE)

rdirichlet(n, alpha)
}
\arguments{
\item{x}{vector of quantiles.}

\item{alpha}{Vector of \code{alpha} parameter values.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}
}
\value{
For \code{ddirichlet}, the kernel density. For \code{rdirichlet}, a
matrix of draws with \code{n} rows and \code{length(alpha)} columns.
}
\description{
A multivariate generalization of the Beta distribution.
}
\author{
Alexey Shiklomanov
}
