% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hector_emissions.R
\name{emissions2ini}
\alias{emissions2ini}
\title{Create Hector INI file from emissions `data.frame`}
\usage{
emissions2ini(emissions, base_ini = system.file("input/hector_rcp45.ini",
  package = "hector"),
  replace_string = "emissions/RCP45_emissions\\\\.csv",
  replace_exo = "exo_emissions.*=0",
  emissions_outfile = tempfile(fileext = ".csv"),
  ini_outfile = tempfile(fileext = ".ini"))
}
\arguments{
\item{emissions}{`data.frame` containing Hector emissions data}

\item{base_ini}{Path to base INI file. Default is Hector RCP 4.5
INI file.}

\item{replace_string}{Regular expression for emissions string to
replace. Default is `"emissions/RCP45_emissions\\.csv"`.}

\item{replace_exo}{Regular expression for exogenous emissions. This
will be deleted because, by default, we pass two values, not a
CSV file. Default = `"exo_emissions.*=0"`}

\item{emissions_outfile}{Target output emissions CSV file. Default
= `tempfile(fileext = ".csv")`.}

\item{ini_outfile}{Target output INI file. Default =
`tempfile(fileext = ".ini")`.}
}
\value{
Path to INI file
}
\description{
Create Hector INI file from emissions `data.frame`
}
\author{
Alexey Shiklomanov
}
