% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ini.R
\name{make_ini}
\alias{make_ini}
\title{Add exogenous CSV data to a Hector INI file}
\usage{
make_ini(exo_csv, ini_in, ini_out)
}
\arguments{
\item{exo_csv}{Path to exogenous emissions CSV file}

\item{ini_in}{Path to baseline Hector INI file}

\item{ini_out}{Target path for output INI file}
}
\value{
\code{ini_out}, invisibly
}
\description{
Add exogenous CSV data to a Hector INI file
}
\author{
Alexey Shiklomanov
}
