% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xml.R
\name{make_xmls}
\alias{make_xmls}
\title{Process GCAM input XML files}
\usage{
make_xmls(hector_ini, hector_xml_in, gcam_xml_in, hector_xml_out,
  gcam_xml_out, gcam_scenario_name = "hector_permafrost")
}
\arguments{
\item{hector_ini}{Path to Hector INI file to use with scenario}

\item{hector_xml_in}{Path to reference Hector XML file to modify}

\item{gcam_xml_in}{Path to reference GCAM XML file to modify}

\item{hector_xml_out}{Path to target Hector XML file}

\item{gcam_xml_out}{Path to target GCAM XML file}

\item{gcam_scenario_name}{Desired GCAM scenario name, as character.
Default = \code{"hector_permafrost"}.}
}
\value{
Named \code{character(2)} containing paths to Hector and GCAM
output files, invisibly.
}
\description{
Process GCAM input XML files
}
\author{
Alexey Shiklomanov
}
