% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mksym}
\alias{mksym}
\title{Create a symbol from a combination of characters and symbols}
\usage{
mksym(...)
}
\arguments{
\item{...}{Characters or symbols. Each of these will be passed to \code{\link[rlang:quo_name]{rlang::quo_name()}}}
}
\value{
A single symbol (see \code{\link[rlang:sym]{rlang::sym()}})
}
\description{
Create a symbol from a combination of characters and symbols
}
\author{
Alexey Shiklomanov
}
