% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{pq}
\alias{pq}
\alias{pred_q}
\title{Helper functions for performing the sensitivity analysis calculations}
\usage{
pq(param)

pred_q(x, i)
}
\arguments{
\item{param}{Unquoted name of a parameter}

\item{x}{Parameters in sensitivity analysis, as a list of symbols}

\item{i}{Index of target parameter for sensitivity calculations}
}
\value{
Both functions return a list of quosures (via
\code{\link[rlang:quos]{rlang::quos()}}), which are then spliced together in the main
\code{\link[=sensitivity_analysis]{sensitivity_analysis()}} function.
}
\description{
These functions heavily leverage \code{rlang}'s quasiquotation mechanism
to build the code that will be used to perform the sensitivity
analysis.
}
\details{
\code{pq} builds the
}
\seealso{
\code{\link[=mksym]{mksym()}} for creating column names, \link[rlang:`!!()`]{rlang::!!()} and \link[rlang:`!!!()`]{rlang::!!!()}
}
