% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_interpolate.R
\name{read_and_interpolate}
\alias{read_and_interpolate}
\title{Read raw emissions data file and interpolate to target years}
\usage{
read_and_interpolate(file, start_year = NULL, end_year = 2100,
  scale = 1, ...)
}
\arguments{
\item{file}{Raw emissions data file}

\item{start_year}{Target start year for interpolation. If \code{NULL}
(default), use smallest year in input data.}

\item{end_year}{Target end year for interpolation. If \code{NULL}
(default), use largest year in input data.}

\item{scale}{Numeric factor by which to multiply values, e.g. to
convert units (default = 1; i.e. no conversion).}

\item{...}{Additional arguments to \code{\link[readr:read_csv]{readr::read_csv()}}.}
}
\value{
\code{data.frame} containing interpolated data, with columns
\code{year} and \code{value}.
}
\description{
Read raw emissions data file and interpolate to target years
}
\author{
Alexey Shiklomanov
}
