% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hector_emissions.R
\name{run_hector_emissions}
\alias{run_hector_emissions}
\title{Run Hector for scenario, given as `data.frame`}
\usage{
run_hector_emissions(emissions, ...)
}
\arguments{
\item{emissions}{`data.frame` containing Hector emissions data}

\item{...}{Arguments passed on to \code{emissions2ini}
\describe{
  \item{emissions}{`data.frame` containing Hector emissions data}
  \item{base_ini}{Path to base INI file. Default is Hector RCP 4.5
INI file.}
  \item{replace_string}{Regular expression for emissions string to
replace. Default is `"emissions/RCP45_emissions\\.csv"`.}
  \item{replace_exo}{Regular expression for exogenous emissions. This
will be deleted because, by default, we pass two values, not a
CSV file. Default = `"exo_emissions.*=0"`}
  \item{emissions_outfile}{Target output emissions CSV file. Default
= `tempfile(fileext = ".csv")`.}
  \item{ini_outfile}{Target output INI file. Default =
`tempfile(fileext = ".ini")`.}
}}
}
\description{
Run Hector for scenario, given as `data.frame`
}
\author{
Alexey Shiklomanov
}
