% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Parameter sensitivity analysis}
\usage{
sensitivity_analysis(df, ...)
}
\arguments{
\item{df}{\code{data.frame} of Hector results.}

\item{...}{Unquoted names of columns describing parameters to be
used in sensitivity analysis.}
}
\value{
\code{data.frame} of sensitivity analysis results. See Details.
}
\description{
Perform a PEcAn-like parameter sensitivity analysis to analyze the
contribution of parameters to predictive uncertainty.
}
\details{
This analysis produces three key metrics:
\itemize{
\item "Coefficient of variation (CV)" describes the relative
uncertainty of the input parameter. It is calculated as the ratio
between the input parameter variance and its median value.
\item "Elasticity" is the normalized sensitivity of the model to a
change in one parameter.
}
}
\author{
Alexey Shiklomanov
}
