% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_biome.R
\name{split_biome}
\alias{split_biome}
\title{Create a Hector core with a split biome}
\usage{
split_biome(biome_name, frac_veg = 0.5, frac_soil = frac_veg,
  frac_detritus = frac_soil, frac_npp_flux0 = frac_veg, rcp = "45")
}
\arguments{
\item{biome_name}{(Character) Name of new biome}

\item{frac_veg}{Fraction of global vegetation C in new biome.
Default is 0.5.}

\item{frac_soil}{Fraction of global soil C in new biome. Default is \code{frac_veg}.}

\item{frac_detritus}{Fraction of global detritus C in new biome.
Default is \code{frac_soil} (which in turn defaults to \code{frac_veg})}

\item{frac_npp_flux0}{Fraction if global initial NPP flux in new
biome. Default is \code{frac_veg}.}

\item{rcp}{Representative carbon pathway (RCP) to use. One of
\code{"26"}, \code{"45"}, \code{"60"}, \code{"85"}}
}
\value{
Hector core with biomes \code{"global"} and \code{biome_name}
}
\description{
Create a Hector core with a split biome
}
\author{
Alexey Shiklomanov
}
