% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{tidy_sensitivity}
\alias{tidy_sensitivity}
\title{Clean up the output of the sensitivity analysis}
\usage{
tidy_sensitivity(sensitivity_out)
}
\arguments{
\item{sensitivity_out}{Output of \code{\link[=sensitivity_analysis]{sensitivity_analysis()}}}
}
\value{
Tidy \code{data.frame} of sensitivity output
}
\description{
Remove extraneous columns, reshape the data to a long format, and
break up the output column names into meaningful pieces.
}
