% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scenario.R
\name{create_scenario}
\alias{create_scenario}
\title{Create a Hector emissions scenario}
\usage{
create_scenario(cc = NULL, ch4_frac = NULL, co2 = NULL, ch4 = NULL,
  base_scenario = "RCP45",
  base_scenario_file = system.file(paste0("input/emissions/",
  base_scenario, "_emissions.csv"), package = "hector"))
}
\arguments{
\item{cc}{`data.frame` of total carbon (C) emissions, with columns
`year` and `value` (emissions in GtC/yr).}

\item{ch4_frac}{Fraction of carbon emitted as methane. Length must
be either 1 or `nrow(cc)`.}

\item{co2}{`data.frame` of CO2 emissions (same structure as `cc`)}

\item{ch4}{`data.frame` of CH4 emissions (same structure as `cc`)}

\item{base_scenario}{Character describing base scenario (default = "RCP45").}

\item{base_scenario_file}{Path to base emissions file (default is
`hector` package RCP 4.5 emissions file).}
}
\value{
`data.frame` containing Hector emissions, with additional
  CO2 and CH4 emissions added to target years.
}
\description{
Create a Hector emissions scenario
}
\author{
Alexey Shiklomanov
}
