% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hector_with_params.R
\name{hector_with_params}
\alias{hector_with_params}
\title{Run Hector with parameters set to specific values}
\usage{
hector_with_params(..., .dots = list(), rcp = "45")
}
\arguments{
\item{...}{Named list of parameter values (e.g. \code{beta = 0.3}).
Names must match Hector parameter names.}

\item{.dots}{A named list of arguments. Provides an alternative
specification to \code{...}}
}
\description{
To specify units, set the \code{unit} attribute of the parameter.
}
\examples{
hector_with_params(beta = 0.5, q10_rh = )
}
\author{
Alexey Shiklomanov
}
