% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hector_emissions.R
\name{run_hector_emissions}
\alias{run_hector_emissions}
\title{Run Hector with custom emissions}
\usage{
run_hector_emissions(rcp, exo_emissions_df = NULL,
  suppresslogging = TRUE, ...)
}
\arguments{
\item{rcp}{Character describing RCP to use as baseline. Must be one
of \code{"26", "45", "60", "85"}.}

\item{exo_emissions_df}{\code{data.frame} containing exogenous emissions
data. Must contain column \code{Date} (year of emissions), and one (or
both) of \code{exo_emissions} (exogenous CO2 emissions) or
\code{exo_ch4_emissions} (for CH4).}

\item{suppresslogging}{(bool) If true, suppress all logging (loglevel is ignored in this case).}

\item{...}{Arguments passed on to \code{hector::newcore}
\describe{
  \item{inifile}{(String) name of the hector input file.}
  \item{loglevel}{(int) minimum message level to output in logs (see \code{\link{loglevels}}).}
  \item{suppresslogging}{(bool) If true, suppress all logging (loglevel is ignored in this case).}
  \item{name}{(string) An optional name to identify the core.}
}}
}
\value{
Hector output, as \code{data.frame} (see \code{\link[hector:fetchvars]{hector::fetchvars()}}).
}
\description{
Run Hector with custom emissions
}
\author{
Alexey Shiklomanov
}
